remove references to `gateway`
