+++ .github/workflows/prod-ci-image.yml
@@ -1,5 +1,4 @@
 name: Build Prod Image CI
-
 on:
   workflow_dispatch:
   pull_request:
@@ -8,28 +7,23 @@ on:
   push:
     branches:
       - "*"
-            
 jobs:
   build:
     runs-on: ubuntu-latest
     steps:
-      -
-        name: Checkout
+      - name: Checkout
         uses: actions/checkout@v3
-      -
-        name: Set up QEMU
+      - name: Set up QEMU
         uses: docker/setup-qemu-action@v2
-      -
-        name: Set up Docker Buildx
+      - name: Set up Docker Buildx
         uses: docker/setup-buildx-action@v2
       - name: Build Prod Image CI
         uses: docker/build-push-action@v3
         with:
           context: .
-          file:    prod/Dockerfile
+          file: prod/Dockerfile
           platforms: linux/amd64
           push: false
           build-args: |
             VERSION=latest
             CARGO_INCREMENTAL=0
-
