+++ lib-xps/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "xps_gateway"
+name = "lib-xps"
 version = "0.1.0"
 edition = "2021"
 resolver = "2"
@@ -21,7 +21,6 @@ ctor.workspace = true
 lib-didethresolver.workspace = true
 gateway-types.workspace = true
 hex.workspace = true
-clap = { version = "4.4.18", features = ["derive"] }
 rand = "0.8.5"
 tokio-stream = { version = "0.1", features = ["net"] }
 registry = { path = "../registry" }
