+++ registry/Cargo.toml
@@ -11,7 +11,7 @@ tracing.workspace = true
 tokio.workspace = true
 async-trait.workspace = true
 ethers = { workspace = true, features = ["ws"] }
-gateway-types.workspace = true
+xps-types.workspace = true
 lib-didethresolver.workspace = true
 rustc-hex.workspace = true
 thiserror.workspace = true
