+++ xps/Cargo.toml
@@ -6,3 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+lib-xps = { path = "../lib-xps" }
+anyhow.workspace = true
+clap = { version = "4.4.18", features = ["derive"] }
